huge organizational rework of the entire streaming client
