+++ tda/streaming/__init__.py
@@ -1,0 +1,4 @@
+from . import client
+from . import exceptions
+
+from .client import StreamClient
