+++ tda/streaming/exceptions.py
@@ -1,0 +1,17 @@
+class UnexpectedResponse(Exception):
+    def __init__(self, response, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.response = response
+
+
+class UnexpectedResponseCode(Exception):
+    def __init__(self, response, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.response = response
+
+
+class UnparsableMessage(Exception):
+    def __init__(self, raw_msg, json_parse_exception, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.raw_msg = raw_msg
+        self.json_parse_exception = json_parse_exception
