+++ tda/streaming/services/ACTIVES_OPTIONS.py
@@ -1,0 +1,4 @@
+from .base import _BaseService
+
+class ACTIVES_OPTIONS(_BaseService): 
+    implemented = False
