+++ tda/streaming/services/ADMIN.py
@@ -1,0 +1,4 @@
+from .base import _BaseService
+
+class ADMIN(_BaseService): 
+    implemented = False
