+++ tda/streaming/services/CHART_HISTORY_FUTURES.py
@@ -1,0 +1,4 @@
+from .base import _BaseService
+
+class CHART_HISTORY_FUTURES(_BaseService): 
+    implemented = False
