+++ tda/streaming/services/FOREX_BOOK.py
@@ -1,0 +1,4 @@
+from .base import _BaseBookService
+
+class FOREX_BOOK(_BaseBookService):
+    pass
