+++ tda/streaming/services/FUTURES_BOOK.py
@@ -1,0 +1,4 @@
+from .base import _BaseBookService
+
+class FUTURES_BOOK(_BaseBookService): 
+    pass
