+++ tda/streaming/services/FUTURES_OPTIONS_BOOK.py
@@ -1,0 +1,4 @@
+from .base import _BaseBookService
+
+class FUTURES_OPTIONS_BOOK(_BaseBookService): 
+    implemented = False
