+++ tda/streaming/services/HEARTBEAT.py
@@ -1,0 +1,4 @@
+from .base import _BaseService
+
+class HEARTBEAT(_BaseService): 
+    implemented = False
