+++ tda/streaming/services/LEVELTWO_FUTURES.py
@@ -1,0 +1,4 @@
+from .base import _BaseService
+
+class LEVELTWO_FUTURES(_BaseService): 
+    implemented = False
