+++ tda/streaming/services/LISTED_BOOK.py
@@ -1,0 +1,4 @@
+from .base import _BaseBookService
+
+class LISTED_BOOK(_BaseBookService):
+    pass
