+++ tda/streaming/services/NEWS_HEADLINE_LIST.py
@@ -1,0 +1,4 @@
+from .base import _BaseService
+
+class NEWS_HEADLINE_LIST(_BaseService): 
+    implemented = False
