+++ tda/streaming/services/NEWS_STORY.py
@@ -1,0 +1,4 @@
+from .base import _BaseService
+
+class NEWS_STORY(_BaseService): 
+    implemented = False
