+++ tda/streaming/services/STREAMER_SERVER.py
@@ -1,0 +1,4 @@
+from .base import _BaseService
+
+class STREAMER_SERVER(_BaseService): 
+    implemented = False
