+++ tda/streaming/services/TIMESALE_EQUITY.py
@@ -1,0 +1,5 @@
+from .base import _BaseService
+from .fields import TimesaleFields
+
+class TIMESALE_EQUITY(_BaseService):
+    Fields = TimesaleFields
