+++ tda/streaming/services/TIMESALE_FOREX.py
@@ -1,0 +1,5 @@
+from .base import _BaseService
+from .fields import TimesaleFields
+
+class TIMESALE_FOREX(_BaseService):
+    Fields = TimesaleFields
