+++ tda/streaming/services/TIMESALE_FUTURES.py
@@ -1,0 +1,5 @@
+from .base import _BaseService
+from .fields import TimesaleFields
+
+class TIMESALE_FUTURES(_BaseService):
+    Fields = TimesaleFields
