+++ tda/streaming/services/TIMESALE_OPTIONS.py
@@ -1,0 +1,5 @@
+from .base import _BaseService
+from .fields import TimesaleFields
+
+class TIMESALE_OPTIONS(_BaseService):
+    Fields = TimesaleFields
