+++ tda/streaming/services/__init__.py
@@ -1,0 +1,72 @@
+from .FUTURES_BOOK import FUTURES_BOOK
+from .CHART_EQUITY import CHART_EQUITY
+from .ADMIN import ADMIN
+from .TIMESALE_EQUITY import TIMESALE_EQUITY
+from .FUTURES_OPTIONS_BOOK import FUTURES_OPTIONS_BOOK
+from .LEVELONE_FOREX import LEVELONE_FOREX
+from .LEVELONE_FUTURES import LEVELONE_FUTURES
+from .ACCT_ACTIVITY import ACCT_ACTIVITY
+from .NASDAQ_BOOK import NASDAQ_BOOK
+from .STREAMER_SERVER import STREAMER_SERVER
+from .LISTED_BOOK import LISTED_BOOK
+from .LEVELTWO_FUTURES import LEVELTWO_FUTURES
+from .NEWS_HEADLINE import NEWS_HEADLINE
+from .QUOTE import QUOTE
+from .OPTIONS_BOOK import OPTIONS_BOOK
+from .TIMESALE_FOREX import TIMESALE_FOREX
+from .NEWS_STORY import NEWS_STORY
+from .ACTIVES_NYSE import ACTIVES_NYSE
+from .LEVELONE_FUTURES_OPTIONS import LEVELONE_FUTURES_OPTIONS
+from .NEWS_HEADLINE_LIST import NEWS_HEADLINE_LIST
+from .TIMESALE_OPTIONS import TIMESALE_OPTIONS
+from .FOREX_BOOK import FOREX_BOOK
+from .ACTIVES_OTCBB import ACTIVES_OTCBB
+from .OPTION import OPTION
+from .CHART_HISTORY_FUTURES import CHART_HISTORY_FUTURES
+from .ACTIVES_OPTIONS import ACTIVES_OPTIONS
+from .CHART_FUTURES import CHART_FUTURES
+from .HEARTBEAT import HEARTBEAT
+from .TIMESALE_FUTURES import TIMESALE_FUTURES
+from .ACTIVES_NASDAQ import ACTIVES_NASDAQ
+
+from .base import _BaseService
+from .fields import *
+
+__service_lookup = {
+    'FUTURES_BOOK': FUTURES_BOOK,
+    'CHART_EQUITY': CHART_EQUITY,
+    'ADMIN': ADMIN,
+    'TIMESALE_EQUITY': TIMESALE_EQUITY,
+    'FUTURES_OPTIONS_BOOK': FUTURES_OPTIONS_BOOK,
+    'LEVELONE_FOREX': LEVELONE_FOREX,
+    'LEVELONE_FUTURES': LEVELONE_FUTURES,
+    'ACCT_ACTIVITY': ACCT_ACTIVITY,
+    'NASDAQ_BOOK': NASDAQ_BOOK,
+    'STREAMER_SERVER': STREAMER_SERVER,
+    'LISTED_BOOK': LISTED_BOOK,
+    'LEVELTWO_FUTURES': LEVELTWO_FUTURES,
+    'NEWS_HEADLINE': NEWS_HEADLINE,
+    'QUOTE': QUOTE,
+    'OPTIONS_BOOK': OPTIONS_BOOK,
+    'TIMESALE_FOREX': TIMESALE_FOREX,
+    'NEWS_STORY': NEWS_STORY,
+    'ACTIVES_NYSE': ACTIVES_NYSE,
+    'LEVELONE_FUTURES_OPTIONS': LEVELONE_FUTURES_OPTIONS,
+    'NEWS_HEADLINE_LIST': NEWS_HEADLINE_LIST,
+    'TIMESALE_OPTIONS': TIMESALE_OPTIONS,
+    'FOREX_BOOK': FOREX_BOOK,
+    'ACTIVES_OTCBB': ACTIVES_OTCBB,
+    'OPTION': OPTION,
+    'CHART_HISTORY_FUTURES': CHART_HISTORY_FUTURES,
+    'ACTIVES_OPTIONS': ACTIVES_OPTIONS,
+    'CHART_FUTURES': CHART_FUTURES,
+    'HEARTBEAT': HEARTBEAT,
+    'TIMESALE_FUTURES': TIMESALE_FUTURES,
+    'ACTIVES_NASDAQ': ACTIVES_NASDAQ
+}
+
+def get_service_class(service_name):
+    return __service_lookup.get(service_name)
+
+def get_service_classes():
+    return list(__service_lookup.values())
