+++ tda/streaming/services/service_list
@@ -1,0 +1,30 @@
+FUTURES_BOOK
+CHART_EQUITY
+ADMIN
+TIMESALE_EQUITY
+FUTURES_OPTIONS_BOOK
+LEVELONE_FOREX
+LEVELONE_FUTURES
+ACCT_ACTIVITY
+NASDAQ_BOOK
+STREAMER_SERVER
+LISTED_BOOK
+LEVELTWO_FUTURES
+NEWS_HEADLINE
+QUOTE
+OPTIONS_BOOK
+TIMESALE_FOREX
+NEWS_STORY
+ACTIVES_NYSE
+LEVELONE_FUTURES_OPTIONS
+NEWS_HEADLINE_LIST
+TIMESALE_OPTIONS
+FOREX_BOOK
+ACTIVES_OTCBB
+OPTION
+CHART_HISTORY_FUTURES
+ACTIVES_OPTIONS
+CHART_FUTURES
+HEARTBEAT
+TIMESALE_FUTURES
+ACTIVES_NASDAQ
